rename class to date
