+++ src/Date.php
@@ -2,7 +2,10 @@
 
 namespace Baka\Support;
 
-class Time
+use DateTime;
+use Exception;
+
+class Date
 {
     /**
      * Calculate how long ago was this time compare to now.
@@ -13,9 +16,13 @@ class Time
      * @param  boolean $timestamp
      * @return string
      */
-    public static function howLongAgo(string $time): string
+    public static function howLongAgo(string $date): string
     {
-        $ts = strtotime($time);
+        if (!self::validate($date)) {
+            throw new Exception('Params must be a valid date Y-m-d H:i:s format');
+        }
+
+        $ts = strtotime($date);
 
         $now = time();
         $diff = $now - $ts;
@@ -40,12 +47,13 @@ class Time
             $text = sprintf(ngettext('%d hr.', '%d hrs.', $hrs), $hrs);
 
             if ($mins > 0) {
-                $text .= ' ' . sprintf(ngettext('%d min.', '%d mins.', $mins), $mins);
+                $text .= ' ' . sprintf(
+                    ngettext('%d min.', '%d mins.', $mins),
+                    $mins
+                );
             }
 
             return sprintf(_('%s ago'), $text);
-
-            return $text;
         } elseif ($diff < $month) {
             $days = floor($diff / $day);
 
@@ -60,5 +68,18 @@ class Time
         } else {
             return date('Y-m-d', $ts);
         }
+    }
+
+    /**
+     * Validate a date given a format
+     *
+     * @param string $date
+     * @param string $format
+     * @return bool
+     */
+    public static function validate(string $date, string $format = 'Y-m-d H:i:s'): bool
+    {
+        $d = DateTime::createFromFormat($format, $date);
+        return $d && $d->format($format) === $date;
     }
 }
